+++ app/include/view/mpv_core.hpp
@@ -118,6 +118,7 @@ public:
 
     // 低画质解码，剔除解码过程中的部分步骤，可以用来节省cpu
     inline static bool LOW_QUALITY = false;
+    inline static bool SUBS_FALLBACK = true;
 
     // 视频缓存（是否使用内存缓存视频，值为缓存的大小，单位MB）
     inline static int INMEMORY_CACHE = 0;