+++ app/src/utils/config.cpp
@@ -59,6 +59,7 @@ std::unordered_map<AppConfig::Item, AppConfig::Option> AppConfig::settingMap = {
     {PLAYER_BOTTOM_BAR, {"player_bottom_bar"}},
     {PLAYER_SEEKING_STEP, {"player_seeking_step", {"5", "10", "15", "30"}, {5, 10, 15, 30}}},
     {PLAYER_LOW_QUALITY, {"player_low_quality"}},
+    {PLAYER_SUBS_FALLBACK, {"player_subs_fallback"}},
     {PLAYER_INMEMORY_CACHE,
         {
             "player_inmemory_cache",
@@ -222,6 +223,7 @@ bool AppConfig::init() {
 #else
     MPVCore::LOW_QUALITY = this->getItem(PLAYER_LOW_QUALITY, false);
 #endif
+    MPVCore::SUBS_FALLBACK = this->getItem(PLAYER_SUBS_FALLBACK, true);
 
     // 初始化是否使用硬件加速
     MPVCore::HARDWARE_DEC = this->getItem(PLAYER_HWDEC, true);